Features are not being ingested due to max age overflow  (#1209)

+++ spark/ingestion/src/main/scala/feast/ingestion/IngestionJobConfig.scala
@@ -95,7 +95,7 @@ case class FeatureTable(
     project: String,
     entities: Seq[Field],
     features: Seq[Field],
-    maxAge: Option[Int] = None
+    maxAge: Option[Long] = None
 )
 
 case class IngestionJobConfig(
